--- a/codegen/generate.go
+++ b/codegen/generate.go
@@ -11,5 +11,6 @@ func GenerateCode(data *Data) error {
 		Data:            data,
 		RegionTags:      true,
 		GeneratedHeader: true,
+		NameForPackage:  data.NameForPackage,
 	})
 }