--- a/internal/imports/prune_test.go
+++ b/internal/imports/prune_test.go
@@ -4,11 +4,12 @@ import (
 	"io/ioutil"
 	"testing"
 
+	"github.com/99designs/gqlgen/internal/code"
 	"github.com/stretchr/testify/require"
 )
 
 func TestPrune(t *testing.T) {
-	b, err := Prune("testdata/unused.go", mustReadFile("testdata/unused.go"))
+	b, err := Prune("testdata/unused.go", mustReadFile("testdata/unused.go"), code.NewNameForPackage(nil))
 	require.NoError(t, err)
 	require.Equal(t, string(mustReadFile("testdata/unused.expected.go")), string(b))
 }