--- a/plugin/stubgen/stubs.go
+++ b/plugin/stubgen/stubs.go
@@ -48,6 +48,7 @@ func (m *Plugin) GenerateCode(data *codegen.Data) error {
 			TypeName: m.typeName,
 		},
 		GeneratedHeader: true,
+		NameForPackage:  data.NameForPackage,
 	})
 }